Get status of promoting VPK files
VitaShell replies with the status code 200, which means that the VPK file was promoted without any issues.
Status code 500 will be replied by VitaShell if it failed promoting.
(see at VitaShell repo in main.c)

Status code 502 gets replied when promoting is not supported by the ftp server you are using.
Also checking for exceptions and reporting all errors to the user. Reporting can be disabled by the addional parameter 'displayErrors' which is by default true

Abort the 1-step install when promoting failed, so it wont upload the game data files.

+++ src/com/soywiz/vitaorganizer/tasks/OneStepToVitaTask.kt
@@ -15,7 +15,10 @@ class OneStepToVitaTask(vitaOrganizer: VitaOrganizer, val vpkFile: VpkFile) : Vi
 		sendPromotingVpkTask.performBase()
 
 		status(Texts.format("PROMOTING_VPK"))
-		PsvitaDevice.promoteVpk(sendPromotingVpkTask.vpkPath)
+		if(!PsvitaDevice.promoteVpk(sendPromotingVpkTask.vpkPath)) {
+			status("Promoting failed! Task aborted!")
+			return
+		}
 		PsvitaDevice.removeFile(sendPromotingVpkTask.vpkPath)
 
 		sendDataTask.performBase()
